Update snake body position and head position
This commit updates the position of each segment of the snake's body and sets the position of the snake's head. The code loop through the snake array and updates each segment's position to be equal to the position of the segment in front of it. The position of the snake's head is then set to the headX and headY variables. Additionally, an HTML span element is added to the boardHTML string for each segment of the snake's body, which will be used to create the snake's appearance on the game board.